+++ hr_employee_department_secondary/README.rst
@@ -36,7 +36,7 @@ The main department can't be in the secondary departments list.
 Bug Tracker
 ===========
 
-Bugs are tracked on `GitHub Issues <https://github.com/OCA/hr/issues>`_.
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/hr/tree/12.0/hr_employee_department_secondary>`_ project on GitHub.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
 `feedback <https://github.com/OCA/hr/issues/new?body=module:%20hr_employee_multidepartment%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
