+++ hr_employee_department_secondary/tests/__init__.py
@@ -1,1 +1,1 @@
-from . import test_hr_employee_multidepartment
+from . import test_hr_employee_department_secondary
